[AIRFLOW-6162] Add back serialization as a module (#6718)
