+++ airflow/models/baseoperator.py
@@ -405,6 +405,11 @@ class BaseOperator(Operator, LoggingMixin):
 
         self.dag = dag or DagContext.get_current_dag()
 
+        # subdag parameter is only set for SubDagOperator.
+        # Setting it to None by default as other Operators do not have that field
+        from airflow.models.dag import DAG
+        self.subdag: Optional[DAG] = None
+
         self._log = logging.getLogger("airflow.task.operators")
 
         # Lineage
